[IMP] web: convert dropdown_accordion_item tests to hoot
closes odoo/odoo#158385

Task-id: 3705027
